requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
django-cors-headers==4.5.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.1
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
regex==2024.9.11
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
